Merge wager code

+++ src/casino/Player.java
@@ -4,23 +4,23 @@ public class Player {
 
 	  private String name;
 	  private String nickName;
-	  private int account;
-	  
-	  Player(String name, String nickName, int account){
+	  private double account;
+
+	  Player(String name, String nickName, double account){
 	    this.name = name;
 	    this.nickName = nickName;
 	    this.account = account;
 	  }
-	  
+
 	  Player(){
 	    name = "Blank";
 	    nickName = "Blank";
 	    account = 0;
 	  }
-	  
+
 	  public void setName(String n){
 	    this.name = n;
-	  } 
+	  }
 	  public String getName(){
 	    return name;
 	  }
@@ -30,10 +30,10 @@ public class Player {
 	  public String getNickName() {
 		  return nickName;
 	  }
-	  public void setAccount(int a){
+	  public void setAccount(double a){
 	    this.account = a;
-	  } 
-	  public int getAccount(){
+	  }
+	  public double getAccount(){
 	    return account;
-	  }  
+	  }
 }
